Add files via upload
Here is an example of test cases and bug reports I made as a part of test design phase for one project. The test object was a form, which contained basic features like: name filed, email field, choose a file button etc. 
It was a fun project and an opportunity to practice writing concise test cases.